Move HostConnectionCounters out into management (it isn't JMX exclusive) and make generic for HTTP as well.
