+++ src/server/src/main/java/io/cassandrareaper/management/HostConnectionCounters.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package io.cassandrareaper.management.jmx;
+package io.cassandrareaper.management;
 
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -34,45 +34,45 @@ public final class HostConnectionCounters {
 
   private final MetricRegistry metricRegistry;
 
-  HostConnectionCounters(MetricRegistry metricRegistry) {
+  public HostConnectionCounters(MetricRegistry metricRegistry) {
     this.metricRegistry = metricRegistry;
   }
 
-  void incrementSuccessfulConnections(String host) {
+  public void incrementSuccessfulConnections(String host) {
     try {
       AtomicInteger successes = successfulConnections.putIfAbsent(host, new AtomicInteger(1));
       if (null != successes && successes.get() <= 20) {
         successes.incrementAndGet();
       }
-      LOG.debug("Host {} has {} successfull connections", host, successes);
+      LOG.debug("Host {} has {} successful connections", host, successes);
       if (null != metricRegistry) {
         metricRegistry
             .counter(
                 MetricRegistry.name(
-                    JmxConnectionFactory.class, "connections", host.replace('.', 'x').replaceAll("[^A-Za-z0-9]", "")))
+                    ICassandraManagementProxy.class, "connections", host.replace('.', 'x').replaceAll("[^A-Za-z0-9]", "")))
             .inc();
       }
     } catch (RuntimeException e) {
-      LOG.warn("Could not increment JMX successfull connections counter for host {}", host, e);
+      LOG.warn("Could not increment successful remote management connections counter for host {}", host, e);
     }
   }
 
-  void decrementSuccessfulConnections(String host) {
+  public void decrementSuccessfulConnections(String host) {
     try {
       AtomicInteger successes = successfulConnections.putIfAbsent(host, new AtomicInteger(-1));
       if (null != successes && successes.get() >= -5) {
         successes.decrementAndGet();
       }
-      LOG.debug("Host {} has {} successfull connections", host, successes);
+      LOG.debug("Host {} has {} successful remote management connections", host, successes);
       if (null != metricRegistry) {
         metricRegistry
             .counter(
                 MetricRegistry.name(
-                    JmxConnectionFactory.class, "connections", host.replace('.', 'x').replaceAll("[^A-Za-z0-9]", "")))
+                    ICassandraManagementProxy.class, "connections", host.replace('.', 'x').replaceAll("[^A-Za-z0-9]", "")))
             .dec();
       }
     } catch (RuntimeException e) {
-      LOG.warn("Could not decrement JMX successfull connections counter for host {}", host, e);
+      LOG.warn("Could not decrement successful remote management connections counter for host {}", host, e);
     }
   }
