+++ src/server/src/test/java/io/cassandrareaper/management/jmx/JmxCustomPortTest.java
@@ -23,6 +23,7 @@ import io.cassandrareaper.ReaperException;
 import io.cassandrareaper.core.Cluster;
 import io.cassandrareaper.core.Node;
 import io.cassandrareaper.crypto.Cryptograph;
+import io.cassandrareaper.management.HostConnectionCounters;
 import io.cassandrareaper.management.ICassandraManagementProxy;
 import io.cassandrareaper.storage.cassandra.CassandraStorageFacade;
